webcam refactored but not tested
